Fix memory access violation in parse code
COM_ParseExt is quite frequently used without calls to
PC_StartParseSession and PC_EndParseSession, which means that the index
'parseDataCount' is -1. This seems to be mostly intended, as no data is
actually stored in parseData in this case. The only parts where it
becomes a problem is when a line counter is trying to be increased for
parseData at index -1. Rather than rewriting a lot of code that is using
these parsing functions (as done in OpenJK for example, see
https://github.com/JACoders/OpenJK/commit/e7b36816) the line counting is
now only done if the 'parseDataCount' index is in range.