Fix trying to print empty strings as loading string
This caused crashes with vid_restart, as the whole scene is attempted to
be rendered if the loading string is empty. As some things are not
properly setup to render the entire scene yet after a vid_restart, this
caused crashes as some forceNames can be empty for some reason. praskOo
had a save in front of the fight versus Rosh where this happened.