Fix broken level inits on too fast systems
Delay clients from joining the server too early. On fast systems, the
player could join the server before NPCs were spawned (see NPC_Spawn).
This could break level init scripts that are triggered by the player,
due to the scripting system not being able to find the NPCs. This
happened to several people on t1_inter for example, where the cinematic
was not started, as the responsible script could not find some NPCs and
errord out.